Trim possible mappings to eliminate supersets
possible mappings will initially include all possible supersets.  We
assume that it will always be better to map to a subset and let that
subset map to the superset itself rather than mapping directly to a
superset. Given this assumption, we can just remove the supersets from
the possible mappings and reduce the search space.

Signed-off-by: Keith Packard <keithp@keithp.com>