Add preprocessing to Lola to allow conditional grammar terms
This adds preprocessing so that portions of the grammar may be elided
if desired. To use, place portions of the grammar inside curly braces,
with the open curly brace followed immediately by a preprocessor symbol:

	foo	: bar
	{ppsym
		| bletch
	}
		;

If -Dppsym is included on the lola command line, the extra production
'| bletch' will be included in the grammar. Otherwise it will not.

Signed-off-by: Keith Packard <keithp@keithp.com>