Strip leading/trailing braces when generating 'unique' action
We compare actions as strings to combine the same action into a single
piece of code. As such, we want to remove any non-significant
characters. Leading/trailing curly braces "should" not matter as the
generated parser code doesn't need them, so removing them for purposes
of comparison may cause more sharing to happen.

Signed-off-by: Keith Packard <keithp@keithp.com>